Refactor of pH Control now passes tests and shows 'B' on idle screen when bubbling.

+++ src/UIState/SeeFreeMemory.cpp
@@ -6,6 +6,6 @@
 void SeeFreeMemory::start() {
   LiquidCrystal_TC::instance()->writeLine(prompt(), 0);
   char buffer[17];
-  snprintf(buffer, sizeof(buffer), "%i bytes", TankControllerLib::instance()->freeMemory());
+  snprintf(buffer, sizeof(buffer), "%i bytes", (int)TankControllerLib::instance()->freeMemory());
   LiquidCrystal_TC::instance()->writeLine(buffer, 1);
 }
